decouple as different resource
