+++ langchain4j-azure-aisearch-spring-boot-starter/src/main/java/dev/langchain4j/azure/aisearch/spring/Properties.java
