+++ langchain4j-azure-aisearch-spring-boot-starter/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+dev.langchain4j.azure.aisearch.spring.AutoConfig
