+++ langchain4j-azure-openai-spring-boot-starter/src/main/java/dev/langchain4j/azure/openai/spring/AutoConfig.java
