+++ langchain4j-azure-openai-spring-boot-starter/src/test/java/dev/langchain4j/azure/openai/AutoConfigIT.java
