--- langchain4j-azure-spring-boot-starter/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-dev.langchain4j.azure.openai.spring.AutoConfig,\
-dev.langchain4j.azure.aisearch.spring.AutoConfig
