--- langchain4j-azure-spring-boot-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,0 @@
-dev.langchain4j.azure.openai.spring.AutoConfig
-dev.langchain4j.azure.aisearch.spring.AutoConfig
